Create outputs module and move relevant files there
