+++ parley/src/outputs/data.rs
@@ -3,7 +3,7 @@
 
 use crate::inline_box::InlineBox;
 use crate::inputs::style::Brush;
-use crate::layout::{Alignment, Glyph, LineMetrics, RunMetrics, Style};
+use crate::outputs::{Alignment, Glyph, LineMetrics, RunMetrics, Style};
 use crate::util::nearly_zero;
 use crate::Font;
 use core::ops::Range;
